+++ app/javascript/components/admin/users/ManageUserRow.jsx
@@ -11,7 +11,7 @@ import Avatar from '../../users/Avatar';
 import Modal from '../../shared/Modal';
 import CreateRoomForm from '../../forms/CreateRoomForm';
 import useCreateServerRoom from '../../../hooks/mutations/admins/users/useCreateServerRoom';
-import DeleteUserForm from '../DeleteUserForm';
+import DeleteUserForm from '../forms/DeleteUserForm';
 
 export default function ManageUserRow({ user }) {
   const mutationWrapper = (args) => useCreateServerRoom({ userID: user.id, ...args });
